newspaperapp/pdf-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
Pillow==10.1.0
img2pdf==0.5.1
gunicorn==21.2.0
